Change the double dtype to float for criteria, menas, stds Fix a bug in initBoxEncoderMetaData

+++ rocAL/rocAL/include/api/rocal_api_meta_data.h
@@ -183,8 +183,8 @@ extern "C" void ROCAL_API_CALL rocalGetSequenceFrameTimestamps(RocalContext roca
 /// \param means [x y w h] mean values for normalization.
 /// \param stds [x y w h] standard deviations for offset normalization.
 /// \param scale Rescales the box and anchor values before the offset is calculated (for example, to return to the absolute values).
-extern "C" void ROCAL_API_CALL rocalBoxEncoder(RocalContext p_context, std::vector<double> &anchors, double criteria,
-                                             std::vector<double>  &means , std::vector<double>  &stds ,  bool offset = false, double scale = 1.0);
+extern "C" void ROCAL_API_CALL rocalBoxEncoder(RocalContext p_context, std::vector<double> &anchors, float criteria,
+                                             std::vector<float>  &means , std::vector<float>  &stds ,  bool offset = false, float scale = 1.0);
 
 /// \param boxes_buf  user's buffer that will be filled with encoded bounding boxes . Its needs to be at least of size batch_size.
 /// \param labels_buf  user's buffer that will be filled with encoded labels . Its needs to be at least of size batch_size.
