+++ rocAL/rocAL/include/meta_data/meta_data_graph.h
@@ -36,7 +36,7 @@ public:
     virtual void process(MetaDataBatch* meta_data, bool segmentation) = 0;
     virtual void update_meta_data(MetaDataBatch* meta_data, decoded_image_info decode_image_info, bool segmentation) = 0;
     virtual void update_random_bbox_meta_data(MetaDataBatch* meta_data, decoded_image_info decoded_image_info,crop_image_info crop_image_info) = 0;
-    virtual void update_box_encoder_meta_data(std::vector<double> *anchors, pMetaDataBatch full_batch_meta_data , double criteria, bool offset , double scale, std::vector<double> &means, std::vector<double> &stds) = 0;
+    virtual void update_box_encoder_meta_data(std::vector<double> *anchors, pMetaDataBatch full_batch_meta_data , float criteria, bool offset , float scale, std::vector<float> &means, std::vector<float> &stds) = 0;
     std::list<std::shared_ptr<MetaNode>> _meta_nodes;
 };
